requirements.txt
requests
rapidfuzz
rich
python-dotenv
google-genai
numpy
fastapi
uvicorn[standard]
websockets
